+++ src/clj/rems/db/test_data.clj
@@ -163,13 +163,13 @@
     (assert (:success result) {:command command :result result})
     (:id result)))
 
-(defn create-dynamic-workflow! [{:keys [actor organization title handlers]
-                                 :as command}]
+(defn create-workflow! [{:keys [actor organization title type handlers]
+                         :as command}]
   (let [result (workflow/create-workflow!
                 {:user-id (or actor "owner")
                  :organization (or organization "abc")
                  :title (or title "")
-                 :type :dynamic
+                 :type (or type :workflow/master)
                  :handlers
                  (or handlers
                      (do (create-user! (get +fake-user-data+ "developer"))
@@ -186,7 +186,7 @@
         result (catalogue/create-catalogue-item!
                 {:resid (or resource-id (create-resource! {}))
                  :form (or form-id (create-form! {}))
-                 :wfid (or workflow-id (create-dynamic-workflow! {}))
+                 :wfid (or workflow-id (create-workflow! {}))
                  :localizations (or localizations {})})]
     (assert (:success result) {:command command :result result})
     (:id result)))
@@ -557,10 +557,21 @@
   (let [approver1 (users :approver1)
         approver2 (users :approver2)
         owner (users :owner)
-        dynamic (create-dynamic-workflow! {:actor owner
-                                           :organization "nbn"
-                                           :title "dynamic workflow"
-                                           :handlers [approver1 approver2]})]
+        default (create-workflow! {:actor owner
+                                   :organization "nbn"
+                                   :title "Default workflow"
+                                   :type :workflow/default
+                                   :handlers [approver1 approver2]})
+        decider (create-workflow! {:actor owner
+                                   :organization "nbn"
+                                   :title "Decider workflow"
+                                   :type :workflow/decider
+                                   :handlers [approver1 approver2]})
+        master (create-workflow! {:actor owner
+                                  :organization "nbn"
+                                  :title "Master workflow"
+                                  :type :workflow/master
+                                  :handlers [approver1 approver2]})]
 
     ;; attach both kinds of licenses to all workflows
     (let [link (create-license! {:actor owner
@@ -576,10 +587,12 @@
                                  :license/text {:en (apply str (repeat 10 "License text in English. "))
                                                 :fi (apply str (repeat 10 "Suomenkielinen lisenssiteksti. "))}})]
       (doseq [licid [link text]]
-        (doseq [wfid [dynamic]]
+        (doseq [wfid [default decider master]]
           (db/create-workflow-license! {:wfid wfid :licid licid}))))
 
-    {:dynamic dynamic}))
+    {:default default
+     :decider decider
+     :master master}))
 
 (defn- create-disabled-applications! [catid applicant approver]
   (create-draft! applicant [catid] "draft with disabled item")
@@ -731,10 +744,10 @@
         handlers [(+fake-users+ :approver1)
                   (+fake-users+ :approver2)]
         owner (+fake-users+ :owner)
-        workflow-id (create-dynamic-workflow! {:actor owner
-                                               :organization "perf"
-                                               :title "Performance tests"
-                                               :handlers handlers})
+        workflow-id (create-workflow! {:actor owner
+                                       :organization "perf"
+                                       :title "Performance tests"
+                                       :handlers handlers})
         form-id (create-form!
                  {:actor owner
                   :form/organization "perf"
@@ -849,31 +862,45 @@
         _ (create-archived-form!)
         workflows (create-workflows! +fake-users+)]
     (create-disabled-license! (+fake-users+ :owner))
-    (let [dynamic (create-catalogue-item! {:title {:en "Dynamic workflow"
-                                                   :fi "Dynaaminen työvuo"}
-                                           :infourl {:en "http://www.google.com"
-                                                     :fi "http://www.google.fi"}
-                                           :resource-id res1
-                                           :form-id form
-                                           :workflow-id (:dynamic workflows)})]
-      (create-applications! dynamic +fake-users+))
+    (create-catalogue-item! {:title {:en "Master workflow"
+                                     :fi "Master-työvuo"}
+                             :infourl {:en "http://www.google.com"
+                                       :fi "http://www.google.fi"}
+                             :resource-id res1
+                             :form-id form
+                             :workflow-id (:master workflows)})
+    (create-catalogue-item! {:title {:en "Decider workflow"
+                                     :fi "Päättäjätyövuo"}
+                             :infourl {:en "http://www.google.com"
+                                       :fi "http://www.google.fi"}
+                             :resource-id res1
+                             :form-id form
+                             :workflow-id (:decider workflows)})
+    (let [catid (create-catalogue-item! {:title {:en "Default workflow"
+                                                 :fi "Oletustyövuo"}
+                                         :infourl {:en "http://www.google.com"
+                                                   :fi "http://www.google.fi"}
+                                         :resource-id res1
+                                         :form-id form
+                                         :workflow-id (:default workflows)})]
+      (create-applications! catid +fake-users+))
     (create-catalogue-item! {:title {:en "Dynamic workflow with extra license"
                                      :fi "Dynaaminen työvuo ylimääräisellä lisenssillä"}
                              :resource-id res-with-extra-license
                              :form-id form
-                             :workflow-id (:dynamic workflows)})
+                             :workflow-id (:default workflows)})
     (let [thlform (create-thl-demo-form! +fake-users+)
           thl-catid (create-catalogue-item! {:title {:en "THL catalogue item"
                                                      :fi "THL katalogi-itemi"}
                                              :resource-id res1
                                              :form-id thlform
-                                             :workflow-id (:dynamic workflows)})]
+                                             :workflow-id (:default workflows)})]
       (create-member-applications! thl-catid (+fake-users+ :applicant1) (+fake-users+ :approver1) [{:userid (+fake-users+ :applicant2)}]))
     (let [dynamic-disabled (create-catalogue-item! {:title {:en "Dynamic workflow (disabled)"
                                                             :fi "Dynaaminen työvuo (pois käytöstä)"}
                                                     :resource-id res1
                                                     :form-id form
-                                                    :workflow-id (:dynamic workflows)})]
+                                                    :workflow-id (:default workflows)})]
       (create-disabled-applications! dynamic-disabled
                                      (+fake-users+ :applicant2)
                                      (+fake-users+ :approver1))
@@ -882,7 +909,7 @@
                                                            :fi "Dynaaminen työvuo (vanhentunut)"}
                                                    :resource-id res1
                                                    :form-id form
-                                                   :workflow-id (:dynamic workflows)})]
+                                                   :workflow-id (:default workflows)})]
       (db/set-catalogue-item-endt! {:id dynamic-expired :end (time/now)}))))
 
 (defn create-demo-data! []
@@ -913,20 +940,20 @@
                                                      :fi "Dynaaminen työvuo"}
                                              :resource-id res1
                                              :form-id form
-                                             :workflow-id (:dynamic workflows)})]
+                                             :workflow-id (:default workflows)})]
         (create-applications! dynamic users))
       (let [thlform (create-thl-demo-form! users)
             thl-catid (create-catalogue-item! {:title {:en "THL catalogue item"
                                                        :fi "THL katalogi-itemi"}
                                                :resource-id res1
                                                :form-id thlform
-                                               :workflow-id (:dynamic workflows)})]
+                                               :workflow-id (:default workflows)})]
         (create-member-applications! thl-catid (users :applicant1) (users :approver1) [{:userid (users :applicant2)}]))
       (let [dynamic-disabled (create-catalogue-item! {:title {:en "Dynamic workflow (disabled)"
                                                               :fi "Dynaaminen työvuo (pois käytöstä)"}
                                                       :resource-id res1
                                                       :form-id form
-                                                      :workflow-id (:dynamic workflows)})]
+                                                      :workflow-id (:default workflows)})]
         (create-disabled-applications! dynamic-disabled
                                        (users :applicant2)
                                        (users :approver1))